Fix computation of max*BufferOverhead on 32 bit platforms
sizeOf (0 :: Int) on 64 bit happens to be 8 which is the same as the
maximum amount of ASCII characters needed to encode any Word32 in
hexadecimal. This doesn't hold on other platforms, so it's better to
hardcode this value.

I've documented the overhead calculation to the best of my knowledge
though I'm not a 100% sure about the precise reasoning. Especially
in the case of maxAfterBufferOverhead, I'm wondering why it isn't
`crlfLength + maxWord32HexLength + crlfLength * 2` which would mean that
a terminating chunk (which is always empty) would fit in the buffer
after the chunk being written at the moment.

This basically has the same effect as the following fix from
bsb-http-chunked and should prevent edge cases where buffers aren't big
enough on 32 bit platforms.
https://github.com/sjakobi/bsb-http-chunked/commit/dde7c9fa33bb6e55b44c5f3e3024215475f64d4c